server/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
PyJWT==2.9.0
pydantic==2.9.2
